requirements.txt
discord>=2.3.2
GitPython>=3.1.42
keyboard>=0.13.5
mss>=9.0.1
numpy>=1.25.2
opencv-python>=4.8.0
Pillow>=10.2.0
PyAutoGUI>=0.9.54
pygame>=2.5.1
pyotp>=2.9.0
pynput>=1.7.6
pyserial>=3.0
pytz>=2023.3
tensorflow>=2.13.0
tensorflow-macos>=2.13.0; sys_platform == "darwin"
tensorflow-metal>=1.0.0; sys_platform == "darwin"
scikit-learn>=1.3.0
matplotlib>=3.7.0
